Move services and CodeFactor improvements
