+++ src/app/config/no-config/no-config.component.scss
@@ -11,7 +11,7 @@
         height: 4vh;
         border-radius: 2vh;
         background-color: #44bd32;
-        width: 0vw;
+        width: 0;
         transition: width .7s ease-in-out;
 
         &-wrapper {
