+++ src/app/files/files.component.scss
@@ -95,7 +95,6 @@
         white-space: nowrap;
         text-overflow: ellipsis;
         display: block;
-        white-space: nowrap;
     }
 
     &__icon {
