+++ src/app/layer-progress/layer-progress.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { LayerProgressService, DisplayLayerProgressAPI } from './layer-progress.service';
+import { LayerProgressService, DisplayLayerProgressAPI } from '../plugin-service/layer-progress.service';
 import { Subscription } from 'rxjs';
 
 @Component({
