+++ src/app/plugin-service/layer-progress.service.ts
